Guest login and Article List on Document Ready
Made guest login that signs in when the page is loaded so that users can
have guest permissions to create/view comments and view/search articles.

Sign out re-signs up the user as guest.

Empty out .article-container div when showArticle is executed so old
articles don't stack with new ones.